WIP portrait definitions
