+++ W3ChampionsStatisticService/Admin/Portraits/PortraitDefinitions.cs
@@ -1,0 +1,14 @@
+﻿using System.Collections.Generic;
+
+namespace W3ChampionsStatisticService.Admin.Portraits
+{
+    public class PortraitDefinitions
+    {
+        public PortraitDefinitions(List<int> _ids)
+        {
+            Ids = _ids;
+        }
+
+        public List<int> Ids { get; set; }
+    }
+}
