+++ WC3ChampionsStatisticService.UnitTests/IntegrationTestBase.cs
@@ -11,8 +11,8 @@ namespace WC3ChampionsStatisticService.UnitTests
 {
     public class IntegrationTestBase
     {
-        // protected readonly MongoClient MongoClient = new MongoClient("mongodb://localhost:27017/");
-        protected readonly MongoClient MongoClient = new MongoClient("mongodb://157.90.1.251:3512/");
+        protected readonly MongoClient MongoClient = new MongoClient("mongodb://localhost:27017/");
+        // protected readonly MongoClient MongoClient = new MongoClient("mongodb://157.90.1.251:3512/");
 
         protected PersonalSettingsProvider personalSettingsProvider;
